Separated Resultpanel code from homepage, navigation process is now different when you go back from settings, with some artificial delay for calculation
